Add Select2 4.x version and rename it select4
This is a bit heavy-handed, but we need a transitional path from v3 to v4 of the select2 library,
and as-written the two versions cannot co-exist, due to a naming conflict in javascript
as well as numerous naming conflicts with css classes.

This solution renames the library to select4 by replacing every string in the js and the css.