APIv4 - Return ID field from getFields when action is create
Early on in APIv4 development, it seemed like a clever idea to hide the 'id' field from getFields
when the action was 'create'. After all, it's not possible to change that field value.
However, it is valid to use it to identify a record in the 'save' action, which uses the same getFields data.
Also, we now have the new "readonly" designation, which is more precise than entirely hiding the field.