Add error hint
This exception is *also* raised, when no membership types have been configured yet, which can be very confusing. It cost me a two hours to figure out the real reason, and I have the nagging suspicion that this had happened to me before. I'm not too happy about the phrasing, but it would be good to have a reference to a potentially incomplete configuration. Thanks.