Container - Use ClassScanner to setup HookListeners. Add test.
Before: While setting up the container-cache, it scans for these patterns:

* BAOs implementing HookInterface
* BAOs implementing EventDispatcherInterface

After: While setting up the container-cache, it scans for these patterns:

* BAOs implementing HookInterface (*unchanged*)
* BAOs implementing EventDispatcherInterface (*unchanged*)
* Extensions for classes (./Civi and ./CRM) implementing HookInterface (*new*)

Comment: For the moment, the `EventDispatcherInterface` is not (on its own)
sufficient to activate auto-registration.  This is because (1) the interface
is quite common in the world-at-large, so we'd prone to over-listening and
(2) we don't control the versioning-cycle, so the dephell is likely worse.

However, if you really want it, I think that it will pick up classes that
implement both `HookInterface` and `EventDispatcherInterface`.