resetNavigation() - Chill out
It appears that this was added as part of 96689db3d7b9276b10db3ee21b4cb9eed9ffc166, which included a couple of flushes, eg

1. After changing "Localization" settings, it flushed navigation and lots of other things.
2. After changing "Navigation", it flushed `navigation` and lots of other things.

The first one feels reasonable. But I'm struggling to understand the second one.

In isolation, it might not be that big of a deal -- but it gets a little silly when combined with ManagedEntities.
If you have several navigation items registered via ManagedEntities, then a "system flush" causes them all to
be written, and each one of them fires a separate/general flush.