APIv4 - Enable getFields to report dynamic FKs if a value is supplied for the entity type
E.g. many tables have a DFK with a pair of `entity_table` and `entity_id` columns;
If you supply a value for `entity_table`, then getFields will output the correct `fk_entity` for the `entity_id` field.

This allows the "Assigned Financial Accounts" display to add the financial type in the page title.