(REF) ContactAutofill, ContactDedupe - Fix warning
There was a warning when flushing:

WARNING: Class Civi\Afform\Behavior\ContactAutofill was expected to have a service definition, but it did not. Perhaps it needs service name.
WARNING: Class Civi\Afform\behavior\ContactDedupe was expected to have a service definition, but it did not. Perhaps it needs service name.