Improve contact sample data
My primary goal here was to add an Organization into the sample data
- although none of the built in message templates really seem to need it
right now.

I also went through & standardised the data in the individual
templates into apiv4 style. These are only used in
unit tests and the MessageAdmin extension at the moment
so if tests pass we should be good to go. It may be that I need
to do a couple of long-overdue token fixes before tests
will pass.

In future I hope to access examples from the unit tests too.