dev/core#3981 Fix Examples to use 0 rather than a possibly-real contact_id
Overview
==================================================
dev/core#3981 Fix Examples to use 0 rather than a possibly-real contact_id

https://lab.civicrm.org/dev/core/-/issues/3981

Before
==============================================
When using the Message Admin UI and trying out tokens those that are not
explicitly defined in the sample data are fetched from the database from
the contact whose id has been arbitrarily used.

After
=======================================================================
The contact_id is set to 0. As a result nothing is retrievable from the databse.

Technical Details
=======================================================================
I suspect this was not possible when the example was first written (the
token code has matured a bit) & a stand in number seemed intuitive - but
I think that is arguable even when there are no negative side
effects (for example if a url were in the letter then linking to
a random contact from it is kinda confusing) - but really loses it's potency
as an argument when there ARE negative side effects

Comments
============================================================================
I'm pretty sure there are some tests I need to adjust for this....