Afform - Fix display of search filter fields in PHP 8.1+
In PHP 8.1 Afform search fields stopped working. This is due to an encoding change which broke the parser.

According to https://www.php.net/manual/en/function.htmlspecialchars.php#refsect1-function.htmlspecialchars-changelog
PHP 8.1 changed the default mode for `htmlspecialchars`, the impact of which is to escape straight quotes
instead of leaving them alone.

This ensures the function works the same on all PHP versions by explicitly setting the mode.