Add fastArray caching to mixinBoot & mixinScan
In monitoring I found this was hitting Redis many times per request - since they
hold an array that is going to hit serialisation performance impacts.

Note I don't address this here as this PR should not make it better or worse -
but we should be flushing these at the same time we flush the metadata cache
during extension enable & disable.