Alias CRM_Financial_BAO_FinancialTypeAccount -> CRM_Financial_BAO_EntityFinancialAccount
This BAO class did not follow the standard naming convention which effectively made it invisible to the API.
Why was named this way in the first place was never explained, I suspect it was simply a mistake.

A BAO class actually did get added before this PR, in 37c608f6e which further added to the confusion.
This consolidates the two classes.

Because it was an "invisible" BAO class, the create and del functions were never used by the API.
For consistency, I've therefore tagged them as @deprecated which keeps them hidden from the API.
We can add business logic back in as needed using hook listeners.