Remove incorrect usages of {localize} in civicrm_data.tpl
The {localize} block was being used in this template in ways that could not plausibly work

- Using the tag in some rows but not others within the same INSERT statement
- Adding it to the rows but not the columns of an INSERT statement

I think the only reason this didn't cause a crash is because this file is only used during installation,
and it's not possible to install CiviCRM with multiple languages - you have to switch to multilingual
post-installation. So these {localize} blocks were never actually used.