CiviUnitTestCase - Display more accurate debug note
If you skimmed, you might say: It prints a console message when it resets
the DB.  That seems like useful information for someone who wants to
understand the overall behavior, so good.

The reality is different.  `setUpBeforeClass()` and `setUp()` both reset the
DB.  `setUpBeforeClass()` does it for every test-class, but it never says so.
`setUp()` does it for the first test-method (and then no others).

So when it showed the message "Installing database", it should have really said:

> I just installed the database a minute ago without telling you.
> Now I'm installing it a second time, and you should know that.
> In the future, I will install twenty more times, but you don't need to know that.

That's a silly thing to say. Better to say when it's happening for realz.