FiveFiftySeven - Switch upgrade step from #25170 (5.57.beta1 => 5.57.0)
This was split apart from 25170 to simplify to some other aspects of the concurrent merges
required for the security-update. Those are done, and now we can proceed.

This will ensure that anyone who ran 5.57 RC will also run the step.  The
step was merged in very, very recently (like 10-20 min ago). For all other
purposes, it's the same as before.