Add new Config function to register templates
I had been upgrading some of the extensions civix to use the
addTemplateDir method (this is already merged to civix
and replaces the more cludgey way that fails post Smarty2).

However, it struck me that we are still calling a
class that extends the Smarty class - which means we have
fewer options than if we call a separate class, with no
Smarty dependency.

The specific challenge currently is that
loading version 3 from an extension 'works' but I have to
directly hack the require_once in the file because the
CRM_Core_Smarty file is loaded by the first class that
attempts to run it's stock civix config. In theory if
the classes didn't do this then the extension's auto-loader
would kick in and we could do a class_exists check
first. I'm sure it's more difficult than that but
I think not involving the Smarty class directly
in our config functions we have more options
to evolve it.

I made the signature an array since that felt more flexible
although only type=Smarty is a thing and
the version is ignored at this stage

Note this would also require a civix update & the goal is to
start phasing this in in preparation for a future option to
use Smarty3

Related https://github.com/civicrm/civicrm-core/pull/25248