CustomField - Fix smarty notices on field create/edit form
The structure of the form was changed back in 2020 to stop using a hierarchical quickform select,
so references to `$form.data_type.value.0` etc are no longer valid.
This was resulting in massive smarty notices.

This removes the offending clauses which were not doing anything as the show/hide is now handled by jQuery.