Allow extension gettext mo files to live in the I18N resource dir
CiviCRM core already supports having the civicrm.mo files in a custom
directory, defined by the CIVICRM_L10N_BASEDIR constant. With this patch,
it will also be possible to have the 'mo' files from extensions in that
directory.

It will be assumed that if an extension is called "foo" then the mo file
will be foo.mo and in the same directory as civicrm.mo.