Increase timeout on imap
About once a day we get an error because the imap server takes longer than the 5
second timeout. This would be largely invisible to most sites as it
would cause a run of fetch_bounces to fail quietly & then the next
one would probably work. For us the impact is we get notified.

I think the impact of increasing to 15 seconds is small enough we
don't need to make it configurable and I doubt that we hit this
timeout any more than anyone else (our imap is within our
own network which is surely more reliably fast than gmail)
so the main impact is that people's fetch_bounces jobs are likely
to result in data getting into civi with less failed runs & hence
a bit quicker