Fix slow queries in snapshot process
SELECT COUNT(*) is slow to run on a large database but SELECT MAX(id) is
quick. The snapshot check only needs to be very rough - so MAX(id) should
be accurate enough.

This does make the message a little inaccurate - but my money says no-one
will notice & it's not worth creating work for the translators to change