Creating a new Drupal user with some kinds of CiviCRM custom fields throws an exception.
To reproduce, add an Alphanumeric Checkbox custom field to the user registration
profile with at east one field option. As an admin, create a new user and check
one of the options for the custom field and save. An exception like
"CRM_Core_Exception: '1' is not a valid option for field custom_1 in
civicrm_api3()" is thrown.

It looks like the $params passed to civicrm_api3() is pretty raw and mainly
derived from $_POST. CRM_Contact_BAO_Contact::createProfileContact() does a
bunch of transformation of the $params before it goes to
CRM_Contact_BAO_Contact::create(). Maybe there's an existing utility function
to translate the raw $_POST values to ones that are API ready, but I haven't
been able to find it.

For now, the easiest fix would be to revert that part of the change in
288eddac95e6c and use "CRM_Contact_BAO_Contact::createProfileContact()" again.