Fix test to test the thing it was written to test
Per https://github.com/civicrm/civicrm-core/pull/23929 the point of the test was
to check that any override coming from parameters or the recurring contribution
record would be respected.

Also see
https://issues.civicrm.org/jira/browse/CRM-16397

However, the test did not explicity change the recurring contribution record amount
field, having been written after a hook was added that was changing the recurring amount
based on the received payment. So it was not testing it's goal - that where they differed
then, for single line item contributions, the recurring record took precedence.

This fixes to specifically alter them to be different, thus testing the override works
and does not rely on the hook (which is likely to change it's behaviour)