Fix in ContributionTest class
This fixes a set up bug. The test deals with a recurring contribution where the
tax is involved. The set up creates the recurring with an amount of 500
and the contribution with an amount of 100. This is invalid but because of
a code change that was outside of what is being tested it scrapes through.

This fixes the test to test a valid 100 for each, followed by the recurring
being changed to an override amount ot 200.