dev/core#4158 Fix incorrect option_group update in paypal recurring
This line was pointed out as being problematic in the issue
https://lab.civicrm.org/dev/core/-/issues/4158 and in a test
I was able to verify it causes issues if the value for Completed
in the contribution status option group is not the same as the
contribution recur option group. However, it appears to
be more of a theoretical / test creatable regression than a real
world one & is not a credible fix for the bug.

I think an rc merge without a stable back port is OK