Fix profile search listings columns are sometimes blank.
The bug is that the query results have the column names formatted like
[location_type_machine_name]-[field_name], like "Work-email", while the
code that handles rendering the row is looking for
[location_type_display_name]-[field_name], like "Office_1-email". I've
adjusted the code that handles rendering the row to instead use the
machine name rather than the display name for the location type.

This only becomes an issue if the location type display name has been
changed from the default so it no longer matches the internal machine
name.