test/phpunit/api/v3 - Fix spooky interaction
This fixes an issue observed in review of 25673. With that patch included,
`JobProcessMailingTest` would still pass... but only in isolation. When
combined with `ExceptionTest`, it would fail.

I tried disabling the entire substance of `ExceptionTest`, and it still
failed. Say what?

The problem was _the order_ in which `ExceptionTest` does its initialization
(`parent::setUp()` and `$this->useTransaction()`). I suppose that order
determines the effectiveness of the transaction.

Most tests call `setup()` then `useTransaction()`. I looked for
outliers which flipped the order -- and found `EntityTagACLTest`
has the same init and the same spooky interaction.

After fixing the outliers, I'm getting the expected results on
`JobProcessMailingTest`.