PaymentProecssor - Add frontend_title & make title required
Clarifies the role of 'title' as "for admins" and adds a new "frontend_title" for end-users.

Updates the PaymentProcessor form/dao to use hooks instead of an overloaded "create" function,
and only allows the title and frontend_title instead of the name to be entered on the form.

The name will be autogenerated from the title in the DAO::writeRecord function.
The frontend_title is required in the database but will be filled from `title` if not supplied during `create`

Updates tests, sql and sample data with required field