Afform - Fix showing default values for EntityRef fields
There was some back-and-forth with this. #25374 removed the change() trigger, which fixed
the recursion bug described in https://lab.civicrm.org/dev/core/-/issues/4083 but broke
setting defaults, prompting #25465 which proposed to change it back.

This fix walks the line by triggering the change event but only when the value has actually changed.
Fixes dev/core#4110