Fix code handling to pre-5.61 for membership block
The work for #26170 concluded that membershipPriceset is always TRUE
at this point. Since we were pushing out a regression fix I left the
variable where it didn't seem to be causing regressions.

However, I think clarifying the variable and
removing the redundant if-else in the rc will set us up well as
1) if there is any further regression we haven't found in
5.61 then people will be able to check on the rc / master
2) if we need to do any further fixes / back-ports the
code will be more synced if we do this in the rc

Note there is some further clean up I feel should be done
folloiwng what we learnt dealing with #26170 - but
I want to let the dust settle so that I'm not creating a
hard-to-backport situation