Minor cleanup on function
The primary goal here was to check that it ALWAYS returns an array. It does,
so I added a type hint.

I also confirmed that
- it is not called from universe
- it is not called with no primaryParticipantID
- it is not called with excludeCancel set for false
so I did some deprecating.

Lastly there was code to handle there being no cancelStatus -
assigning a variable cancelID of 0 as a fallback. But, it
was not being used - so I fixed it to cast-to-int.