APIv4 - Fix html encoding of rich-text fields
In ece8de2d2 this was fixed but only for APIv3, and with no unit test.
The previous fix also did not cover fields using "TextArea" as their input type,
even though they are allowed to store HTML.
This fixes for APIv4 and v3 and adds a test. and adds a test.