Add is_active to link spec
In the process of adding weights to the links it
occurred to me that it might be cleaner in some cases for
hooks to fiddle with 'is_active' rather than setting &
unsetting.

An example would be that instead of excluding
a link that our core code thinks should not be available
we could include it with is_active = FALSE &
a hook could enable it.

I figured if it I add it now then while adding weights I could
bank it in too