Declare paid event as example data, call from test
I'm looking to create some set up helper classes in
the Civi\Test space, to also be available to extensions
& to standardise some of the set up routines.

I figured the best place to put the ExampleData
was in the structure created for that.

The ExampleData structure is lightly used by
the WorkflowMessage::render() function for personal
entities (ie Contact,Contribution) but it
is not used (& probably won't be) for
configuration entities. However, the examples
are intended to be multi-purpose so I think it is the right location.

Note the trait that builds on it will have to have
a lot more logic to build price sets etc than the exampleData
space really supports