PriceSet forms - remove reference to undeclared _contributionAmount property
I don't find this value anywhere in our universe outside this
function - which implies it should be not be a property - especially for
php 8,x reasons. In fact it really just is a case of checking if field_name
is contribution_amount at the right places. I didn't include it in the place where
it would read if field_name = x AND field_name <> y....