regen.sh - Preserve the default component-extensions
The issue is that `civicrm_setting[enable_components]` and
`civicrm_extension` need to be sync'd (as of 5.63+).

But there's a structural asymmetry - when dealing with an empty/default
arrangement, "settings"/"components" will inherit a default activation-list;
but "extensions" start with an empty activation-list.

With this patch, we retain the local-active list of component-extensions.
Hopefully, that matches the default setting...