APIv4 - Enable UNION and UNION ALL
Supports UNION and UNION ALL with global limiting/sorting/having. Per SQL UNION rules,
names do not need to match (columns in the UNIONed tables will be aliased to
match the column names of the first table) but the number of columns and the data
types do need to match.