Add All imports menu item for users with Administer QUeues
This adds the (only just added) search display for 'all imports' to the menu, but
only for users with 'Administer Queues' - for other users they would only see
their own imports, so it would duplicate 'my imports'.

This is useful for site admins etc trying to understand issues their users might be having
with imports.

Note the extension needs to be enabled