Fix e-notices in contribution_offline_receipt text by removing always-empty field
On looking to fix this enotice through the form it turned out the
contribution form does not have a field for this - so the formValues
will never hold receipt_text.

Hence this PR removes it from the template.

Other forms do seem to have this & I will separately add userText or similar to the model