REF - Auto-generate links for CRM_Core_Page_Basic
Instead of hard-coding an array of links in every page, we can auto-generate the list
based on the 'paths' metadata in the DAO. As a POC this converts one page (ACLs).
Future conversions are simply a matter of ensuring 'paths' is defined then deleting the
links function and $_links variable.