SearchKit - Reuse helpers from CRM_Core_Action
This simplifies SearchKit code by reusing the new helpers added in cf957fc
It adds an additional helper for fetching the icon associated with an action.
Also fixes an action name in the CustomField DAO which did not match the name
of a CRM_Core_Action (renamed 'move' to 'detach' which is the convention other entities use).