Security Checks - Update error report for unsupported HTTPD.
Some security checks don't work on PHP 7.4's built-in HTTPD -- they cause
the system hang for >1min (*while internal HTTP calls timeout*). The
checks run whenever the admin logs in -- so having it hang is fairly
annoying (*for the developer who uses PHP built-in HTTPD*).

This only affects local development, so it's not critical to run the checks.

We'll skip the affected tests and show a simple (quick) error instead.