Minor cleanup on handling of financial type
The frontend & back office contribution forms both pass financialType as an
option to processRecurringContribution, no other place calls this function.

The value is taken from params & then overwritten from the object.

This simplifies by setting in the calling function as an array.

The function should be split up & returned to the forms that call it (&
refactored into nothingness) but I think it makes sense to
stop passing financialType as a first step