Fold preProcess into run
there are properties that appear to be being set solely to pass variables from preProcess to
run.

preProcess is called from run and that is the only way it is accessed.

There are some forms that use the pageRun hook in the universe but this has no impact on them