ActionMapping - Clean up obtuse methods of fetching a MappingInterface class
Changed calls to getMappings to use getMapping if they only needed a single value.
In some cases they could be entirely removed because they weren't using the retrieved class for anything.
Deprecated the use of 'filters' param in getMappings as it was redundant with simply calling getMapping with an id.