getEntitySpecificJoins sometimes returns NULL, triggering deprecation warning for trim() in php 8.1
Discovered this in the wild where an api3 Contact.getcount call for a particular smartgroup
was causing this. The smartgroup otherwise works fine.

Looking at the code, it's clear that getEntitySpecificJoins is supposed to return a string,
but in the case of civicrm_activity_contact it was returning from another function which may
return NULL. I have coalesced NULL to empty string.