Clarify, params passed to preApproval
I dug into what happens to the params that is passed
to preApproval after that function ends, The answer is - nothing.
It uses them to interact with the gateway and from the
returned result session parameters are set, and
a redirect might happen. But the process calling it
is on a path to exit after that and there is no expectation
of anything being changed when it passes-by-reference.
The equivalent function on the event form passes a local variable
which is not used again.

In order to make that clearer I switched to a local variable on the
contribution form too.

 Along with that
I moved the setting of keys only used in the if clause into the if clause
and switched to use frontend_title in the description